// src/app/interfaces/detallePlan.ts
export interface DetallePlan{
  id: string;
  alimentacion: string;
  hospedaje: string;
  transporte: string;
  traslados: string;
  valor: string;
  cantidad_noches: string;
  cantidad_dias: string;
  fecha_creacion: string;
  fecha_modificacion: string;
  usu_creador: string;
  usu_modificador: string;
  estado: string;
  id_plan: string;
  id_dest: string;
}